implemented a reusable splineAreaSeriesChartComponent
thermometerChartComponent will be deprecated to be replaced by the reusable splineAreaSeriesChartComponent.
Created Interface for the new reusable chart component so it is easier to know what values are required to
reuse the component in other parts of the aplication. Added dummy data for testing the chart component to be removed
when needed. Replaced deprecated thermometerChartCompnent from dashboardCharts component.