Added external social media icons to nav bar
Added new font inorder to use brand font icons like instagram, twitter etc.
Styled brand icons in nav bar. Exported method that resolves Label centering bug on android devices
to a directive. This was done to avoid the repetative initialization code that was needed before inorder to fix this bug on Labels that needed to be centered.
Added "fab" new scss class for icons that are brands.