Work in progress setting up charts to recieve data from backend
Added external webview for media link, fixed bug that casued
hamburger menu icon to diapear when on webview,
Added the abilty to hit http for devleopment and
testing backend with app. set up charts service and component
to consume backend data but is work in progress.
Added title property to splineAreaChart.